+++ people-and-teams/profilecard/src/components/User/ProfileCardTrigger.tsx
@@ -11,6 +11,7 @@ import {
   ProfilecardProps,
   ProfileCardTriggerProps,
   ProfileCardTriggerState,
+  TeamCentralReportingLinesData,
 } from '../../types';
 import { DELAY_MS_HIDE, DELAY_MS_SHOW } from '../../util/config';
 
@@ -75,6 +76,7 @@ class ProfilecardTrigger extends React.PureComponent<
     hasError: false,
     error: null,
     data: null,
+    reportingLinesData: undefined,
   };
 
   componentDidMount() {
@@ -123,10 +125,14 @@ class ProfilecardTrigger extends React.PureComponent<
         data: null,
       },
       () => {
-        this.props.resourceClient
-          .getProfile(cloudId || '', userId)
+        const requests = Promise.all([
+          this.props.resourceClient.getProfile(cloudId || '', userId),
+          this.props.resourceClient.getReportingLines(userId),
+        ]);
+
+        requests
           .then(
-            (res) => this.handleClientSuccess(res),
+            (res) => this.handleClientSuccess(res[0], res[1]),
             (err) => this.handleClientError(err),
           )
           .catch((err) => this.handleClientError(err));
@@ -134,7 +140,10 @@ class ProfilecardTrigger extends React.PureComponent<
     );
   };
 
-  handleClientSuccess(res: ProfileCardClientData) {
+  handleClientSuccess(
+    profileData: ProfileCardClientData,
+    reportingLinesData: TeamCentralReportingLinesData,
+  ) {
     if (!this._isMounted) {
       return;
     }
@@ -142,7 +151,8 @@ class ProfilecardTrigger extends React.PureComponent<
     this.setState({
       isLoading: false,
       hasError: false,
-      data: res,
+      data: profileData,
+      reportingLinesData,
     });
   }
 
@@ -167,6 +177,8 @@ class ProfilecardTrigger extends React.PureComponent<
       clientFetchProfile: this.clientFetchProfile,
       analytics: this.props.analytics,
       ...this.state.data,
+      reportingLines: this.state.reportingLinesData,
+      onReportingLinesClick: this.props.onReportingLinesClick,
     };
 
     const wrapperProps =